+++ huynh38/step1/parser/SymbolTable.h
@@ -9,11 +9,11 @@ using namespace std;
 class SymbolTable {
 public:
 	static SymbolTable* getInstance();
-	void insertEntry(string key, pair<double, double>);
+	void insertEntry(string key, pair<int, int>);
 	pair<double, double> getEntry(std::string key);
 	static int getSize();
 	static int getLen();
-	vector<map<string, pair<double, double>>> SymbolTableList;
+	vector<map<string, pair<int, int>>> SymbolTableList;
 	void setScope(int _scope);
 	int getScope();
 	//static int getData(string key);
